dyninst: address review feedback on swiss map index PR
- Remove dead wyhash fallthrough comment (grantseltzer)
- Widen ValInSlotOffset from uint8 to uint16 across IR, compiler,
  bytecode, and BPF to support map values with fields past offset 255
- Extract swissMapOps() helper to deduplicate condition/expression
  handler code in generate.go (piob-io)
- Add HeaderByteSize field to SwissMapSetupOp so the BPF opcode is
  self-contained instead of reading buf_offset_1 from the preceding
  DereferenceOp (piob-io)
- Convert bare type assertions to ,ok idiom in resolveSwissMapIndex
  (piob-io)
- Fix bool key validation in validateSwissMapKeyLiteral to accept
  bool literals for base-type map keys
- Add test cases: boundary key lengths (16/17), zero-value key,
  nil pointer value (ExprStatusNilDeref), bool keys, and large
  value struct (field at offset 256)

+++ pkg/dyninst/ebpf/context.h
@@ -146,7 +146,7 @@ typedef struct stack_machine {
     uint8_t ctrl_offset;
     uint8_t slots_offset;
     uint8_t key_in_slot_offset;
-    uint8_t val_in_slot_offset;
+    uint16_t val_in_slot_offset;
     uint8_t key_byte_size;
     uint8_t is_string_key;
